Move Address validation into Address class
- get rid of usages of eip55 address where possible
